fix exit code of processes in si_waitpid on windows
    GetExitCodeProcess returns the exit code in a DWORD, which is an
    unsigned integer, however main functions usually return a signed int, so
    we need put in an explicit type cast.

+++ src/c/unixsys.d
@@ -152,7 +152,7 @@ si_waitpid(cl_object pid, cl_object wait)
     code = ECL_NIL;
   } else {
     status = @':exited';
-    code = ecl_make_fixnum(exitcode);
+    code = ecl_make_fixnum((int)exitcode);
     pid->foreign.data = NULL;
     CloseHandle(*hProcess);
   }
